Commit - Syncing latest changes before lunch 2018-06-07 with the rep.

+++ Sjuklöner/Controllers/CareCompaniesController.cs
@@ -10,6 +10,7 @@ using Sjuklöner.Models;
 using Microsoft.AspNet.Identity;
 using Sjuklöner.Viewmodels;
 using static Sjuklöner.Viewmodels.OmbudIndexVM;
+using System.Text.RegularExpressions;
 
 namespace Sjuklöner.Controllers
 {
@@ -105,6 +106,8 @@ namespace Sjuklöner.Controllers
         public ActionResult Edit([Bind(Include = "CareCompanyId,CareCompany,SelectedCollectiveAgreementId,CollectiveAgreement")] CareCompanyEditVM careCompanyEditVM, string submitButton)
         {
             //ModelState.Remove(nameof(CareCompany.CollectiveAgreementSpecName));
+            careCompanyEditVM.SelectedCollectiveAgreementId = careCompanyEditVM.CareCompany.SelectedCollectiveAgreementId;
+            List<SelectListItem> collectiveAgreements = new List<SelectListItem>();
             if (ModelState.IsValid)
             {
                 if (submitButton == "Spara")
@@ -122,10 +125,24 @@ namespace Sjuklöner.Controllers
                     careCompany.SelectedCollectiveAgreementId = careCompanyEditVM.CareCompany.SelectedCollectiveAgreementId;
                     db.Entry(careCompany).State = EntityState.Modified;
                     db.SaveChanges();
-                    return View(careCompanyEditVM);
+
+                    //collectiveAgreements = db.CollectiveAgreementHeaders.ToList().ConvertAll(c => new SelectListItem
+                    //{
+                    //    Value = $"{c.Id}",
+                    //    Text = c.Name
+                    //});
+                    //careCompanyEditVM.CollectiveAgreement = new SelectList(collectiveAgreements, "Value", "Text");
+                    //return View(careCompanyEditVM);
                 }
                 return RedirectToAction("Index", "Claims");
             }
+
+            collectiveAgreements = db.CollectiveAgreementHeaders.ToList().ConvertAll(c => new SelectListItem
+            {
+                Value = $"{c.Id}",
+                Text = c.Name
+            });
+            careCompanyEditVM.CollectiveAgreement = new SelectList(collectiveAgreements, "Value", "Text");
             return View();
         }
 
@@ -282,8 +299,77 @@ namespace Sjuklöner.Controllers
         //public ActionResult EditOmbud([Bind(Include = "Id,FirstName,LastName,LastLogon,CareCompanyId,SSN,Email,EmailConfirmed,PasswordHash,SecurityStamp,PhoneNumber,PhoneNumberConfirmed,TwoFactorEnabled,LockoutEndDateUtc,LockoutEnabled,AccessFailedCount,UserName")] ApplicationUser applicationUser)
         public ActionResult EditOmbud([Bind(Include = "Id,FirstName,LastName,CareCompanyId,CareCompanyName,SSN,Email,PhoneNumber")] OmbudEditVM ombudEditVM, string submitButton)
         {
+            var currentId = User.Identity.GetUserId();
+            ApplicationUser currentUser = db.Users.Where(u => u.Id == currentId).FirstOrDefault();
+
+            bool errorFound = false;
+            //Check that the SSN has the correct format
+            if (!string.IsNullOrWhiteSpace(ombudEditVM.SSN))
+            {
+                ombudEditVM.SSN = ombudEditVM.SSN.Trim();
+                Regex regex = new Regex(@"^([1-9][0-9]{3})(((0[13578]|1[02])(0[1-9]|[12][0-9]|3[01]))|((0[469]|11)(0[1-9]|[12][0-9]|30))|(02(0[1-9]|[12][0-9])))[-]?\d{4}$");
+                Match match = regex.Match(ombudEditVM.SSN);
+                if (!match.Success)
+                {
+                    ModelState.AddModelError("SSN", "Ej giltigt personnummer. Formaten YYYYMMDD-NNNN och YYYYMMDDNNNN är giltiga.");
+                    errorFound = true;
+                }
+            }
+            else
+            {
+                errorFound = true;
+            }
+
+            //Check that the ombud is born in the 20th or 21st century
+            if (!errorFound)
+            {
+                if (int.Parse(ombudEditVM.SSN.Substring(0, 2)) != 19 && int.Parse(ombudEditVM.SSN.Substring(0, 2)) != 20)
+                {
+                    ModelState.AddModelError("SSN", "Ombudet måste vara fött på 1900- eller 2000-talet.");
+                    errorFound = true;
+                }
+            }
+
+            //Check that the ombud is at least 18 years old and was not born in the future:-)
+            if (!errorFound)
+            {
+                DateTime ombudBirthday = new DateTime(int.Parse(ombudEditVM.SSN.Substring(0, 4)), int.Parse(ombudEditVM.SSN.Substring(4, 2)), int.Parse(ombudEditVM.SSN.Substring(6, 2)));
+                if (ombudBirthday.Date > DateTime.Now.Date)
+                {
+                    ModelState.AddModelError("SSN", "Födelsedatumet får inte vara senare än idag.");
+                    errorFound = true;
+                }
+                else if (ombudBirthday > DateTime.Now.AddYears(-18))
+                {
+                    ModelState.AddModelError("SSN", "Ombudet måste vara minst 18 år.");
+                    errorFound = true;
+                }
+            }
+
+            //Check if there is an ombud with the same SSN already in the company. The same ombud is allowed in another company.
+            if (!errorFound)
+            {
+                var twinOmbud = db.Users.Where(u => u.SSN == ombudEditVM.SSN).Where(u => u.Id != ombudEditVM.Id).FirstOrDefault();
+                if (twinOmbud != null && twinOmbud.CareCompanyId == currentUser.CareCompanyId)
+                {
+                    ModelState.AddModelError("SSN", "Det finns redan ett ombud med detta personnummer");
+                    errorFound = true;
+                }
+            }
+
+            if (!errorFound)
+            {
+                if (ombudEditVM.SSN.Length == 12)
+                {
+                    ombudEditVM.SSN = ombudEditVM.SSN.Insert(8, "-");
+                }
+            }
+
             if (submitButton == "Spara")
             {
+                var possibleTwin = db.Users.Where(u => u.Email == ombudEditVM.Email).FirstOrDefault();
+                if (possibleTwin != null && possibleTwin.Id != ombudEditVM.Id)
+                    ModelState.AddModelError("Email", "Det finns redan en användare med den emailadressen");
                 if (ModelState.IsValid)
                 {
                     var editedOmbud = db.Users.Where(u => u.Id == ombudEditVM.Id).FirstOrDefault();
@@ -331,9 +417,15 @@ namespace Sjuklöner.Controllers
         {
             if (submitButton == "Bekräfta")
             {
+                var myId = User.Identity.GetUserId();
+                var me = db.Users.Where(u => u.Id == myId).FirstOrDefault();
                 ApplicationUser applicationUser = db.Users.Find(id);
-                db.Users.Remove(applicationUser);
-                db.SaveChanges();
+                if(applicationUser != me && applicationUser.CareCompanyId == me.CareCompanyId)
+                {
+                    db.Users.Remove(applicationUser);
+                    db.SaveChanges();
+                }
+
             }
             return RedirectToAction("IndexOmbud");
         }
